refactor : 도메인 별 클래스 분리 (#582)
