+++ src/main/java/com/clubber/ClubberServer/domain/review/mapper/ReviewMapper.java
@@ -1,20 +1,11 @@
 package com.clubber.ClubberServer.domain.review.mapper;
 
-import com.clubber.ClubberServer.domain.admin.domain.Admin;
-import com.clubber.ClubberServer.domain.admin.dto.GetAdminPendingReviewsSliceResponse;
-import com.clubber.ClubberServer.domain.admin.dto.GetAdminsPendingReviews;
-import com.clubber.ClubberServer.domain.admin.dto.GetAdminsReviewsResponse;
-import com.clubber.ClubberServer.domain.admin.dto.AdminReviewResponse;
-import com.clubber.ClubberServer.domain.club.domain.Club;
 import com.clubber.ClubberServer.domain.review.domain.Review;
 import com.clubber.ClubberServer.domain.review.dto.ClubReviewResponse;
 import com.clubber.ClubberServer.domain.review.dto.CreateClubReviewResponse;
 import com.clubber.ClubberServer.domain.review.dto.GetClubReviewsPageResponse;
 import com.clubber.ClubberServer.domain.review.dto.GetClubReviewsSliceResponse;
 import com.clubber.ClubberServer.domain.review.util.ReviewUtil;
-import com.clubber.ClubberServer.domain.user.domain.User;
-import com.clubber.ClubberServer.domain.user.dto.GetUserReviewsResponse;
-import com.clubber.ClubberServer.domain.user.dto.UserReviewResponse;
 import com.clubber.ClubberServer.global.common.page.PageResponse;
 import com.clubber.ClubberServer.global.common.slice.SliceResponse;
 import com.clubber.ClubberServer.global.util.SliceUtil;
@@ -68,61 +59,6 @@ public class ReviewMapper {
 				}
 			)
 			.collect(Collectors.toList());
-	}
-
-	// 리뷰 조회 (일반 회원)
-	public GetUserReviewsResponse getGetUserReviewResponse(User user, List<Review> reviews) {
-		List<UserReviewResponse> userReviewResponse = getUserReviewResponse(
-			reviews);
-		return GetUserReviewsResponse.of(user, userReviewResponse);
-	}
-
-	private static List<UserReviewResponse> getUserReviewResponse(
-		List<Review> reviews) {
-		return reviews.stream()
-			.map(
-				review -> {
-					Set<String> keywords = ReviewUtil.extractKeywords(review);
-					return UserReviewResponse.of(review, keywords);
-				}
-			)
-			.collect(Collectors.toList());
-	}
-
-	// 리뷰 조회 (관리자)
-	public GetAdminsReviewsResponse getGetAdminReviewsResponse(
-		Admin admin, Club club, Page<Review> reviews) {
-		PageResponse<AdminReviewResponse> adminsReviewDetailsPageResponse = getAdminsReviewResponse(
-			reviews);
-		return GetAdminsReviewsResponse.of(admin, club, adminsReviewDetailsPageResponse);
-	}
-
-	private static PageResponse<AdminReviewResponse> getAdminsReviewResponse(
-		Page<Review> reviewPages) {
-		Page<AdminReviewResponse> getAdminReviewsPageResponse = reviewPages.map(
-			review -> {
-				Set<String> keywords = ReviewUtil.extractKeywords(review);
-				return AdminReviewResponse.of(review, keywords);
-			});
-		return PageResponse.of(getAdminReviewsPageResponse);
-	}
-
-	//대기 상태 리뷰 조회 (관리자)
-	public List<GetAdminsPendingReviews> getGetAdminPendingReviewList(List<Review> reviews){
-		return reviews.stream()
-			.map(GetAdminsPendingReviews::from)
-			.collect(Collectors.toList());
-	}
-
-	//대기 상태 리뷰 조회 (관리자, No-offset)
-	public GetAdminPendingReviewsSliceResponse getGetAdminPendingReviewSliceResponse(
-		List<Review> reviews, Pageable pageable){
-		List<GetAdminsPendingReviews> getAdminPendingReviewList = getGetAdminPendingReviewList(
-			reviews);
-		SliceResponse<GetAdminsPendingReviews> getAdminsPendingReviewsSliceResponse = SliceUtil.valueOf(
-			getAdminPendingReviewList, pageable);
-		Long lastReviewId = ReviewUtil.getLastReviewId(reviews, pageable);
-		return GetAdminPendingReviewsSliceResponse.of(getAdminsPendingReviewsSliceResponse, lastReviewId);
 	}
 
 	// 리뷰 작성
