+++ src/main/java/com/clubber/ClubberServer/domain/user/service/UserService.java
@@ -4,12 +4,12 @@ import com.clubber.ClubberServer.domain.favorite.domain.Favorite;
 import com.clubber.ClubberServer.domain.favorite.dto.GetFavoriteDetailsResponse;
 import com.clubber.ClubberServer.domain.favorite.repository.FavoriteRepository;
 import com.clubber.ClubberServer.domain.review.domain.Review;
-import com.clubber.ClubberServer.domain.review.mapper.ReviewMapper;
 import com.clubber.ClubberServer.domain.user.dto.GetUserReviewsResponse;
 import com.clubber.ClubberServer.domain.review.repository.ReviewRepository;
 import com.clubber.ClubberServer.domain.user.domain.User;
 import com.clubber.ClubberServer.domain.user.dto.GetUserFavoritesResponse;
 import com.clubber.ClubberServer.domain.user.dto.GetUserProfileResponse;
+import com.clubber.ClubberServer.domain.user.mapper.UserMapper;
 import com.clubber.ClubberServer.global.common.page.PageResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -29,7 +29,7 @@ public class UserService {
 
 	private final UserReadService userReadService;
 
-	private final ReviewMapper reviewMapper;
+	private final UserMapper userMapper;
 
 	public GetUserProfileResponse getUserProfile() {
 		User user = userReadService.getUser();
@@ -45,7 +45,7 @@ public class UserService {
 	public GetUserReviewsResponse getUserReviews() {
 		User user = userReadService.getUser();
 		List<Review> reviews = reviewRepository.queryReviewByUserOrderByIdDesc(user);
-		return reviewMapper.getGetUserReviewResponse(user, reviews);
+		return userMapper.getGetUserReviewResponse(user, reviews);
 	}
 
 	public PageResponse<GetFavoriteDetailsResponse> getUserFavoritesPagination(Pageable pageable) {
